Fix an error in response to a request for track ID
There was an error in the code that handles a request for sector ID.
This request isn't used by the Brother 930 machine, but is used by
the KH940, so the code path has not been tested until recently.

Also, added a handler for keyboard interrupt to prevent ending with
backtrace vomit.